@@ -28,12 +28,11 @@
     "GPL-2.0-or-later"
   ],
   "replace": {
-    "bugfix/patchcache": "self.version",
     "typo3-ter/patchcache": "self.version"
   },
   "require": {
-    "php": ">=7.0.0,<=7.99",
-    "typo3/cms-core": ">=9.5,<9.6"
+    "php": "^7.2",
+    "typo3/cms-core": "^9.5.0"
   },
   "autoload": {
     "psr-4": {
